requirements.txt
google-api-python-client==2.111.0
langdetect==1.0.9
whisper==1.1.10
pytube==15.0.0
youtube-transcript-api==0.6.1
psycopg2-binary==2.9.9